avoiding to delete non empty courses

+++ src/app/components/cards/course-card/course-card.component.ts
@@ -3,7 +3,7 @@ import { Router } from '@angular/router';
 import { Course } from 'src/app/models/course.model';
 import { Student } from 'src/app/models/student.model';
 import { CourseService } from 'src/app/services/course-service/course.service';
-import {MatDialog} from '@angular/material/dialog';
+import { MatDialog } from '@angular/material/dialog';
 import { StudentCardComponent } from '../student-card/student-card.component';
 import { StudentInfoComponent } from '../../pop-ups/student-info/student-info.component';
 import { StudentService } from 'src/app/services/student-service/student.service';
@@ -19,7 +19,7 @@ export class CourseCardComponent {
   constructor(private router: Router,
     private service: CourseService,
     private serviceStudent: StudentService,
-    private dialog:MatDialog) { }
+    private dialog: MatDialog) { }
 
   showStudents: boolean = false;
 
@@ -33,13 +33,20 @@ export class CourseCardComponent {
   }
 
   onDelete(): void {
-    if (confirm(`Do you want to delete the course with ID: ${this.course?.courseId}?`)) {
-      if (this.course) {
-        this.service.deleteByID(<string>this.course?.courseId).subscribe((answer) => {
-          //Fix the server response
-          console.log(answer)
-          alert(`Course with ID: ${this.course?.courseId} has been deleted!`)
-        })
+    if (this.course.studentList) {
+      if (this.course.studentList.length > 0) {
+        alert('Please remove the students of the course')
+      }
+      else {
+        if (confirm(`Do you want to delete the course with ID: ${this.course?.courseId}?`)) {
+          if (this.course) {
+            this.service.deleteByID(<string>this.course?.courseId).subscribe((answer) => {
+              //Fix the server response
+              console.log(answer)
+              alert(`Course with ID: ${this.course?.courseId} has been deleted!`)
+            })
+          }
+        }
       }
     }
   }
@@ -55,11 +62,11 @@ export class CourseCardComponent {
   }
 
   onShowStudents(): void {
-    this.showStudents= !this.showStudents;
+    this.showStudents = !this.showStudents;
   }
 
-  openDialog(student:Student) {
-    const dialogRef = this.dialog.open(StudentInfoComponent,{data: student});
+  openDialog(student: Student) {
+    const dialogRef = this.dialog.open(StudentInfoComponent, { data: student });
 
     dialogRef.afterClosed().subscribe(result => {
       console.log(`Dialog result: ${result}`);
@@ -67,18 +74,18 @@ export class CourseCardComponent {
   }
 
   openAllStudentsDialog() {
-    const dialogRef = this.dialog.open(AvailableStudentsComponent, {data: this.course});
+    const dialogRef = this.dialog.open(AvailableStudentsComponent, { data: this.course });
 
     dialogRef.afterClosed().subscribe(result => {
       console.log(`Dialog result: ${result}`);
     });
   }
 
-  onUnregisterStudent(student:Student): void {
-    this.serviceStudent.update(<string>student.studentId,student
-      ).subscribe((answer)=>{
-        console.log(answer)
-        alert(`Student with ID: ${answer.data.studentId} has been updated!`)
-      })
+  onUnregisterStudent(student: Student): void {
+    this.serviceStudent.update(<string>student.studentId, student
+    ).subscribe((answer) => {
+      console.log(answer)
+      alert(`Student with ID: ${answer.data.studentId} has been updated!`)
+    })
   }
 }
